websitebackend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
firebase-admin
python-dotenv
google-genai
requests
cloudinary
